+++ lib/processor/process_field_test.go
@@ -137,6 +137,102 @@ func TestProcessFieldString(t *testing.T) {
 	}
 }
 
+func TestProcessFieldCodecs(t *testing.T) {
+	type testCase struct {
+		name   string
+		codec  string
+		input  string
+		output string
+	}
+	tests := []testCase{
+		{
+			name:   "string 1",
+			codec:  "string",
+			input:  `{"target":"foobar"}`,
+			output: `{"target":"foobar"}`,
+		},
+		{
+			name:   "int 1",
+			codec:  "int",
+			input:  `{"target":"5"}`,
+			output: `{"target":5}`,
+		},
+		{
+			name:   "float 1",
+			codec:  "float",
+			input:  `{"target":"5.67"}`,
+			output: `{"target":5.67}`,
+		},
+		{
+			name:   "bool 1",
+			codec:  "bool",
+			input:  `{"target":"true"}`,
+			output: `{"target":true}`,
+		},
+		{
+			name:   "bool 2",
+			codec:  "bool",
+			input:  `{"target":"false"}`,
+			output: `{"target":false}`,
+		},
+		{
+			name:   "object 1",
+			codec:  "object",
+			input:  `{"target":"{}"}`,
+			output: `{"target":{}}`,
+		},
+		{
+			name:   "object 2",
+			codec:  "object",
+			input:  `{"target":"{\"foo\":{\"bar\":\"baz\"}}"}`,
+			output: `{"target":{"foo":{"bar":"baz"}}}`,
+		},
+		{
+			name:   "array 1",
+			codec:  "array",
+			input:  `{"target":"[]"}`,
+			output: `{"target":[]}`,
+		},
+		{
+			name:   "array 2",
+			codec:  "array",
+			input:  `{"target":"[1,2,\"foo\"]"}`,
+			output: `{"target":[1,2,"foo"]}`,
+		},
+	}
+
+	procConf := NewConfig()
+	procConf.Type = "noop"
+
+	for _, test := range tests {
+		t.Run(test.name, func(tt *testing.T) {
+			conf := NewConfig()
+			conf.Type = "process_field"
+			conf.ProcessField.Path = "target"
+			conf.ProcessField.ResultType = test.codec
+			conf.ProcessField.Processors = append(conf.ProcessField.Processors, procConf)
+
+			c, err := New(conf, nil, log.Noop(), metrics.Noop())
+			if err != nil {
+				tt.Fatal(err)
+			}
+
+			exp := [][]byte{
+				[]byte(test.output),
+			}
+			msg, res := c.ProcessMessage(message.New([][]byte{
+				[]byte(test.input),
+			}))
+			if res != nil {
+				tt.Error(res.Error())
+			}
+			if act := message.GetAllBytes(msg[0]); !reflect.DeepEqual(act, exp) {
+				tt.Errorf("Wrong result: %s != %s", act, exp)
+			}
+		})
+	}
+}
+
 func TestProcessFieldBadProc(t *testing.T) {
 	conf := NewConfig()
 	conf.Type = "process_field"